fix(test): move bootstrap test with set_network_id to their own mod - there was a race condition when multiple tests were ran from the same   file using the set_network_id fn. This fn would set thread level   global values which caused the race condition.
