+++ .github/workflows/merge.yml
@@ -120,6 +120,9 @@ jobs:
 
       - uses: Swatinem/rust-cache@v2
 
+      - name: Set ANT_LOG environment variable
+        run: echo "ANT_LOG=all" >> $GITHUB_ENV
+
       - name: Run autonomi tests
         timeout-minutes: 25
         run: cargo test --release --package autonomi --lib
