+++ ant-bootstrap/tests/contacts.rs
@@ -11,7 +11,6 @@ use ant_bootstrap::{
     ContactsFetcher, InitialPeersConfig, ANT_PEERS_ENV,
 };
 use ant_logging::LogBuilder;
-use ant_protocol::version::set_network_id;
 use color_eyre::Result;
 use libp2p::{Multiaddr, PeerId};
 use std::{sync::atomic::Ordering, time::SystemTime};
@@ -131,52 +130,6 @@ async fn test_network_contacts_retries() -> Result<()> {
     let addrs = fetcher.fetch_bootstrap_addresses().await?;
 
     assert!(!addrs.is_empty(), "Should succeed after retries");
-
-    Ok(())
-}
-
-#[tokio::test]
-async fn test_multiple_network_contacts() -> Result<()> {
-    set_network_id(100); // to disable fetching mainnet contacts
-    let _guard = LogBuilder::init_single_threaded_tokio_test();
-    // Start mock servers
-    let mock_server1 = MockServer::start().await;
-    let mock_server2 = MockServer::start().await;
-
-    // Set up endpoints with different peers
-    Mock::given(method("GET"))
-        .and(path("/contacts"))
-        .respond_with(ResponseTemplate::new(200).set_body_string(
-            "/ip4/127.0.0.1/udp/8080/quic-v1/p2p/12D3KooWRBhwfeP2Y4TCx1SM6s9rUoHhR5STiGwxBhgFRcw3UERE"
-        ))
-        .mount(&mock_server1)
-        .await;
-
-    Mock::given(method("GET"))
-        .and(path("/contacts"))
-        .respond_with(ResponseTemplate::new(200).set_body_string(
-            "/ip4/127.0.0.2/udp/8081/quic-v1/p2p/12D3KooWD2aV1f3qkhggzEFaJ24CEFYkSdZF5RKoMLpU6CwExYV5"
-        ))
-        .mount(&mock_server2)
-        .await;
-
-    // Set up CLI args with both endpoints
-    let args = InitialPeersConfig {
-        first: false,
-        addrs: vec![],
-        network_contacts_url: vec![
-            format!("{}/contacts", mock_server1.uri()),
-            format!("{}/contacts", mock_server2.uri()),
-        ],
-        local: false,
-        ignore_cache: true,
-        bootstrap_cache_dir: None,
-    };
-
-    // Should fetch from both endpoints
-    let addrs = args.get_bootstrap_addr(None, None).await?;
-
-    assert_eq!(addrs.len(), 2, "Should fetch from both endpoints");
 
     Ok(())
 }
